register r2 namespace
