+++ packages/wrangler/src/index.ts
@@ -58,7 +58,6 @@ import { APIError, formatMessage, ParseError } from "./parse";
 import { pipelines } from "./pipelines";
 import { pubSubCommands } from "./pubsub/pubsub-commands";
 import { queues } from "./queues/cli/commands";
-import { r2 } from "./r2";
 import { secret, secretBulkHandler, secretBulkOptions } from "./secret";
 import {
 	addBreadcrumb,
@@ -508,9 +507,7 @@ export function createCLIParser(argv: string[]) {
 	});
 
 	// r2
-	wrangler.command("r2", "📦 Manage R2 buckets & objects", (r2Yargs) => {
-		return r2(r2Yargs, subHelp);
-	});
+	register.registerNamespace("r2");
 
 	// d1
 	wrangler.command("d1", `🗄  Manage Workers D1 databases`, (d1Yargs) => {
